+++ app/src/main/java/com/tr4n/moviedb/SplashActivity.kt
@@ -1,0 +1,19 @@
+package com.tr4n.moviedb
+
+import android.content.Intent
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import android.os.Handler
+import android.os.Looper
+
+class SplashActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        Handler(Looper.getMainLooper()).postDelayed({
+            val mainIntent = Intent(this, MainActivity::class.java)
+            startActivity(mainIntent)
+            finish()
+        }, 2000)
+    }
+}
