+++ app/src/main/res/drawable/splash_background.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list
+    xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:drawable="@drawable/splashscreen" />
+    <item>
+        <bitmap android:gravity="center"
+            android:src="@drawable/popcorn" />
+    </item>
+</layer-list>
