+++ app/src/main/res/layout/fragment_favorite.xml
