+++ app/src/main/res/values/dimens.xml
@@ -1,5 +1,4 @@
 <resources>
     <!-- Default screen margins, per the Android Design guidelines. -->
-    <dimen name="activity_horizontal_margin">16dp</dimen>
-    <dimen name="activity_vertical_margin">16dp</dimen>
+    <dimen name="dp_16">16dp</dimen>
 </resources>
