+++ app/src/main/res/values/strings.xml
@@ -1,7 +1,8 @@
 <resources>
     <string name="app_name">Movie DB</string>
-    <string name="title_activity_home">HomeActivity</string>
+    <string name="title_activity_main">MainActivity</string>
     <string name="title_home">Home</string>
     <string name="title_search">Search</string>
     <string name="title_favorite">Favorite</string>
+    <string name="what_do_you">What do you want to watch?</string>
 </resources>
